TNL-5578 Implement DateUtils, prototype instance
The annotation.target.src was assigned to a function instead of a string
because the missing brackets.

  #   test_lms_courseware:ProctoredExamTest.test_timed_exam_flow
  #  test_lms_matlab_problem:MatlabProblemTest.test_run_code
  #  test_studio_course_info.UsersCanAddUpdatesTest.test_delete_course_update
  #  test_account_settings.AccountSettingsPageTest.test_full_name_field

The interface for custom management commands changed in Django 1.8 which
caused the `compute_grades` command to stop working. In order to fix it
I added an `add_arguments` method for specifying the accepted arguments
and updated the command to allow for passing a list of courses to run
the `offline_grade_calculation` against.